% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg.annotation.R
\docType{data}
\name{cpg.annotation}
\alias{cpg.annotation}
\title{cpg.annotation}
\format{
An object of class \code{data.frame} with 485512 rows and 22 columns.
}
\usage{
cpg.annotation
}
\description{
CpG annotations derived from the \href{https://www.bioconductor.org/packages/release/data/annotation/html/IlluminaHumanMethylation450kanno.ilmn12.hg19.html}{IlluminaHumanMethylation450kanno.ilmn12.hg19} bioconductor R package.
See ./inst/cpg.annotation.R for how the annotation data was derived.
}
\keyword{datasets}
