% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{Example CpG methylation dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 93031 columns.
}
\usage{
example.data
}
\description{
Example dataset of TCGA-PRAD patients with prostate cancer DNA methylation data.  The columns are CpG methylation beta-values.
}
\keyword{datasets}
