% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.models.R
\docType{data}
\name{example.models}
\alias{example.models}
\title{example.models}
\format{
An object of class \code{list} of length 2.
}
\usage{
example.models
}
\description{
Example models used for testing/developing the R package. For all models, see \link{all.models}.
}
\keyword{datasets}
