% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.predict.R
\name{features.predict}
\alias{features.predict}
\title{features.predict}
\usage{
features.predict(gene.methy.data, models)
}
\arguments{
\item{gene.methy.data}{A data frame with gene-level methylation data, created by \link{gene.methylation}.  Patients are rows and columns are genes.}

\item{models}{A list of models used to predict features from gene-level methylation data.  The models should come from \code{data('all.models')}.}
}
\description{
Main function used to predict various clinical and molecular features from gene-level methylation data in prostate cancer patients.
}
\examples{
data('example.models');

### example gene-level methylation data
data('example.data.gene.methy');
# note this dataset is derived from the following commands:
# data('example.data');
# example.data.gene.methy <- gene.methylation(example.data);
check <- validate.gene.methy.data(example.data.gene.methy, example.models);
stopifnot(check$val.passed);

features <- features.predict(example.data.gene.methy, example.models);
features;
}
