% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.subtypes.R
\name{predict.subtypes}
\alias{predict.subtypes}
\title{predict.subtypes}
\usage{
\method{predict}{subtypes}(methy.data, prop.missing.cutoff = 0.3, impute.using.all.cpgs = TRUE)
}
\arguments{
\item{methy.data}{A data.frame with patients as rows (rownames give patient ids) and column names give CpG ids.}

\item{prop.missing.cutoff}{The maximum proportion of missing values allowed for each required CpG. KNN imputation is used to impute missing values.}

\item{impute.using.all.cpgs}{TRUE/FALSE indicating whether to impute missing values using all CpGs in \code{methy.data} or only the CpGs required by \link{subtype.model}.  When TRUE, imputation will be slower and use more memory, but should be more accurate.}
}
\description{
Assign patients to four prostate cancer DNA methylation subtypes
}
\examples{
data('subtype.model');

### example CpG data
data('example.data');

subtypes <- predict.subtypes(example.data);
head(subtypes);
}
