% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype.model.R
\docType{data}
\name{subtype.model}
\alias{subtype.model}
\title{subtype.model}
\format{
An object of class \code{pamrtrained} of length 23.
}
\usage{
subtype.model
}
\description{
\href{https://cran.r-project.org/web/packages/pamr/index.html}{pamr} model used for assigning new patients to four prostate cancer DNA methylation subtypes.
}
\keyword{datasets}
