% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype.model.rf.R
\docType{data}
\name{subtype.model.rf}
\alias{subtype.model.rf}
\title{Random Forest model for predicting methylation subtype}
\format{
An object of class \code{rfsrc} (inherits from \code{grow}, \code{class}) of length 51.
}
\usage{
subtype.model.rf
}
\description{
\href{https://cran.r-project.org/web/packages/randomForestSRC/index.html}{randomForestSRC} model used for assigning new patients to four prostate cancer DNA methylation subtypes. Note \code{\link{subtype.model.pamr}} requires all 5,486 subtype-defining CpGs to be measured, whereas \code{subtype.model.rf} can handle missing CpGs through imputation (although ideally you should have as many of the CpGs as possible).
}
\keyword{datasets}
