% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.subtype.model.cpgs.R
\name{validate.subtype.model.cpgs}
\alias{validate.subtype.model.cpgs}
\title{Validate input data for estimate.subtypes()}
\usage{
validate.subtype.model.cpgs(methy.data, prop.missing.cutoff = 0.3)
}
\arguments{
\item{methy.data}{A data.frame with patients as rows (rownames give patient ids) and column names give CpG ids.}

\item{prop.missing.cutoff}{The maximum proportion of missing values allowed for each required CpG. KNN imputation is used to impute missing values.}
}
\value{
\itemize{
\item \code{val.passed} a logical indicating whether the data passed validation
\item \code{check$required.cpgs} a vector of CpG ids that are required for predicting the subtypes
\item \code{missing.cpgs} a vector of CpG ids that are required but completely missing in the data
\item \code{required.cpgs.with.high.missing} a vector of CpG ids that are required and have a proportion of missing values greater than \code{prop.missing.cutoff}
}
}
\description{
Check whether \code{methy.data} contains all CpGs required by \link{subtype.model} for assigning patients to four prostate cancer DNA methylation subtypes.
}
\examples{
data('example.data');
check <- validate.subtype.model.cpgs(example.data);
stopifnot(check$val.passed);

# CpGs required to fit each model:
#check$required.cpgs;

# CpGs that are required but completely missing in your data:
#check$missing.cpgs;

# CpGs that are required and have a proportion of missing values greater than `prop.missing.cutoff`
#check$required.cpgs.with.high.missing;
}
