% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.models.R
\docType{data}
\name{all.models}
\alias{all.models}
\title{all.models}
\format{
An object of class \code{list} of length 2.
}
\usage{
all.models
}
\description{
Models fit using \link[caret]{train} for predicting various clinical and molecular features of patients diagnosed with prostate cancer.
}
\keyword{datasets}
