% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.models.R
\docType{data}
\name{all.models}
\alias{all.models}
\title{Models for predicting clinical and molecular features}
\format{
An object of class \code{list} of length 14.
}
\usage{
all.models
}
\description{
\href{https://cran.r-project.org/web/packages/randomForest/index.html}{Random forest} and \href{https://cran.r-project.org/web/packages/glmnet/index.html}{glmnet} models for predicting various clinical and molecular features of patients diagnosed with prostate cancer.
}
\keyword{datasets}
