% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.gene.methy.R
\docType{data}
\name{example.data.gene.methy}
\alias{example.data.gene.methy}
\title{Example gene methylation dataset}
\format{
An object of class \code{data.frame} with 2 rows and 13622 columns.
}
\usage{
example.data.gene.methy
}
\description{
Example dataset of TCGA-PRAD patients with gene-level prostate cancer DNA methylation data.  The columns contain the median beta-value among CpG islands in the gene promoter region.
See \link{gene.methylation} for an example showing how to calculate gene-level methylation data.
}
\keyword{datasets}
