% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.methylation.R
\name{gene.methylation}
\alias{gene.methylation}
\title{gene.methylation}
\usage{
gene.methylation(methy, print.progress = TRUE)
}
\arguments{
\item{methy}{methylation dataset where rownames give patient ids and columns use CpG ids}

\item{print.progress}{TRUE/FALSE to show progress bar}
}
\description{
Calculate gene-level methylation for a dataset containing CpGs from the Illumina 450k methylation array.
Gene-level methylation is calculated as the median beta-value among CpG islands in the gene promoter region.
}
\examples{
#data(example.data);
#example.data.gene.methy <- gene.methylation(example.data);
}
