% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.features.R
\name{predict.features}
\alias{predict.features}
\title{predict.features}
\usage{
\method{predict}{features}(gene.methy.data, models, prop.missing.cutoff = 0.3)
}
\arguments{
\item{gene.methy.data}{A data frame with gene-level methylation data, created by \link{gene.methylation}.  Patients are rows and columns are genes.}

\item{models}{A list of models used to predict features from gene-level methylation data.  The models should come from \code{data('all.models')}.}

\item{prop.missing.cutoff}{The maximum proportion of missing values allowed for each required gene.}
}
\description{
Main function used to predict various clinical and molecular features from gene-level methylation data in prostate cancer patients.
}
\examples{
data('all.models');

### example gene-level methylation data
data('example.data.gene.methy');
# note this dataset is derived from the following commands:
# data('example.data');
# example.data.gene.methy <- gene.methylation(example.data);
check <- validate.gene.methy.data(example.data.gene.methy, all.models);
stopifnot(check$val.passed);

features <- predict.features(example.data.gene.methy, all.models);
features;
}
